remove commented code and address feedback
